[forge] Default to port forwarding when using operator resize command
Currently the default for forge is not to have direct access to k8s, so the command would fail to connect to nodes even though they are set up successfully. That makes port forward not really optional, and should be the default unless you're specifically configured network access which most people will not.

This inverts the flag and allows you to not port forward if you have direct connection.

Test Plan:

Run the resize command locally and ensure cluster works

```
$ cargo run -p forge-cli -- operator resize --namespace forge-banana --num-validators 10 --validator-image-tag 880bc32441e94ba39f8b2e93c2a1e1b7a0f5dd2d --testnet-image-tag 880bc32441e94ba39f8b2e93c2a1e1b7a0f5dd2d
```